% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_defillama_chain_hist_tvl.R
\name{get_defillama_chain_hist_tvl}
\alias{get_defillama_chain_hist_tvl}
\title{Chain History Data}
\usage{
get_defillama_chain_hist_tvl(name)
}
\arguments{
\item{name}{A string representing the unique name (name) of the protocol.
This name can be obtained from the result of the \code{\link{get_defillama_chain_tvl}} function.}
}
\value{
A `data.table` object containing historical TVL data of the specified chain
The table includes columns for date, TVL in USD. It is sorted by date.
}
\description{
Fetches historical TVL data for a specified Blockchain in USD.
The data is returned as a table sorted by date.
}
\examples{
# total tvl by date
chain_tvl = get_defillama_chain_hist_tvl( "Ethereum" )
tail( chain_tvl )
}
