% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_defillama_chain_tvl.R
\name{get_defillama_chain_tvl}
\alias{get_defillama_chain_tvl}
\title{Chain TVL Data}
\usage{
get_defillama_chain_tvl()
}
\value{
A `data.table` containing the TVL data with columns: name, symbol, tvl, gecko_id, cmcId, and chainId.
The data is sorted by TVL in descending order.
}
\description{
This function retrieves TVL data for various chains
from a specified API and formats it into a data.table for analysis.
}
\examples{
chain_tvl = get_defillama_chain_tvl()
head( chain_tvl )
}
