% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_defillama_protocol_tvl.R
\name{get_defillama_protocol_tvl}
\alias{get_defillama_protocol_tvl}
\title{Protocol TVL Data}
\usage{
get_defillama_protocol_tvl(is_cex_include = F)
}
\arguments{
\item{is_cex_include}{Logical; if FALSE (default), data for CEX
will be excluded from the results. Set to TRUE to include CEX data.}
}
\value{
A `data.table` containing TVL data with columns: symbol, tvl, slug, name, chain, and category.
The data is sorted by TVL in descending order.
}
\description{
This function retrieves Total Value Locked (TVL) data for various protocols
from a specified API and formats it into a table for analysis.
}
\examples{
protocol_tvl = get_defillama_protocol_tvl()
head( protocol_tvl )
}
