% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_defillama_stablecoin_data.R
\name{get_defillama_stablecoin_data}
\alias{get_defillama_stablecoin_data}
\title{Stablecoins Data}
\usage{
get_defillama_stablecoin_data()
}
\value{
A `data.table` with stablecoin information, sorted by market capitalization.
}
\description{
Retrieves stablecoin data, including details like name, symbol, price, and market cap.
The data is returned as a sorted table by market cap.
}
\examples{
stablecoin_data = get_defillama_stablecoin_data()
head( stablecoin_data )
}
