% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_defillama_stablecoin_hist_data.R
\name{get_defillama_stablecoin_hist_data}
\alias{get_defillama_stablecoin_hist_data}
\title{Stablecoins History Data}
\usage{
get_defillama_stablecoin_hist_data(stablecoin_id)
}
\arguments{
\item{stablecoin_id}{An integer representing the ID of the stablecoin. This ID can be obtained
from the result of the \code{\link{get_defillama_stablecoin_data}} function.}
}
\value{
A `data.table` object containing historical data of the specified stablecoin.
The table is sorted by date.
}
\description{
Retrieves stablecoin data, including details like name, symbol, price, and market cap.
The data is returned as a sorted table.
}
\examples{
\donttest{
stablecoin_hist_data = get_defillama_stablecoin_hist_data( 5 )
head( stablecoin_hist_data )
}
}
