% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protocol_tvl.R
\name{get_protocol_tvl}
\alias{get_protocol_tvl}
\title{Fetch TVL Data from API}
\usage{
get_protocol_tvl()
}
\value{
A data.table containing TVL data with columns: symbol, tvl, slug, name, chain, and category.
}
\description{
This function retrieves Total Value Locked (TVL) data for various protocols
from a specified API and formats it into a data.table for analysis.
}
\examples{
tvl = get_protocol_tvl()
}
