% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamar.R
\name{llama_default_params}
\alias{llama_default_params}
\title{Fetch a copy of the default model parameters}
\usage{
llama_default_params()
}
\value{
named list of values
}
\description{
\describe{
  \item{n_ctx       = 512   }{text context}
  \item{gpu_layers  = 0     }{number of layers to store in VRAM}
  \item{seed        = -1    }{RNG seed, -1 for random}
  \item{f16_kv      = true  }{use fp16 for KV cache}
  \item{logits_all  = false }{the llama_eval() call computes all logits, not just the last one}
  \item{vocab_only  = false }{only load the vocabulary, no weights}
  \item{use_mmap    = true  }{use mmap if possible}
  \item{use_mlock   = false }{force system to keep model in RAM}
  \item{embedding   = false }{embedding mode only}
}
}
