% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamar.R
\name{llama_init}
\alias{llama_init}
\title{Initialise llama.cpp by loading a model}
\usage{
llama_init(
  model_path = "/Users/mike/projectsdata/llama.cpp/ggml-vic7b-q5_0.bin",
  params = list()
)
}
\arguments{
\item{model_path}{path to llama.cpp model.}

\item{params}{parameters for the execution of llama.  See \code{llama_default_params()}
for details.  Note: You possibly don't want/need to play with any of 
these unless you really know what you're doing.}
}
\value{
llama context (\code{ctx})
}
\description{
Initialise llama.cpp by loading a model
}
